LayoutTests/imported/w3c/web-platform-tests/websockets/back-forward-cache-with-closed-websocket-connection-ccns.tentative.window-expected.txt


FAIL Testing BFCache support for page with closed WebSocket connection and "Cache-Control: no-store" header. assert_equals: document unexpectedly BFCached expected (undefined) undefined but got (boolean) true
